[optimization][FM][FFM] code refactoring
move breeze/mllib vector conversion to utils
