+++ src/main/scala/com/intel/imllib/optimization/AdaptiveUpdater.scala
@@ -16,34 +16,13 @@
  */
 package com.intel.imllib.optimization
 
-import breeze.linalg.{DenseVector => BDV, SparseVector => BSV, Vector => BV}
-
 import scala.math._
 import org.apache.spark.annotation.DeveloperApi
-import org.apache.spark.mllib.linalg.{DenseVector, SparseVector, Vector}
+import org.apache.spark.mllib.linalg.Vector
 
+import com.intel.imllib.util.vectorUtils._
 trait AdaUpdater extends Serializable {
   val epsilon = 1e-8
-
-  def toBreeze(mllibVec: Vector): BV[Double] = new BDV[Double](mllibVec.toDense.values)
-  def fromBreeze(breezeVector: BV[Double]): Vector = {
-    breezeVector match {
-      case v: BDV[Double] =>
-        if (v.offset == 0 && v.stride == 1 && v.length == v.data.length) {
-          new DenseVector(v.data)
-        } else {
-          new DenseVector(v.toArray) // Can't use underlying array directly, so make a new one
-        }
-      case v: BSV[Double] =>
-        if (v.index.length == v.used) {
-          new SparseVector(v.length, v.index, v.data)
-        } else {
-          new SparseVector(v.length, v.index.slice(0, v.used), v.data.slice(0, v.used))
-        }
-      case v: BV[_] =>
-        sys.error("Unsupported Breeze vector type: " + v.getClass.getName)
-    }
-  }
 }
 /**
  * :: DeveloperApi ::
